// src/types/Users.ts
export type GithubUser = {
    login: string;
    id: number;
    node_id: string;
    avatar_url: string;
    gravatar_id: string;
    url: string;
    html_url: string;
    followers_url: string;
    following_url: string;
    gists_url: string;
    starred_url: string;
    subscriptions_url: string;
    organizations_url: string;
    repos_url: string;
    events_url: string;
    received_events_url: string;
    type: string;
    user_view_type: string;
    site_admin: boolean;
    company?: string;
    email?: string;
    followers?: number;
    following?: 173;
    public_repos?: number;
    public_gists?: 78;
    created_at?: string;
    updated_at?: string;
    bio?: string;
  };